Name + Focus Alignment
