+++ src/styles/optionAlternateTextCompact/base.js
@@ -1,0 +1,7 @@
+const optionAlternateTextCompact = {
+    flexShrink: 1,
+    flexGrow: 1,
+    flexBasis: 'auto',
+};
+
+export default optionAlternateTextCompact;
