+++ src/styles/optionAlternateTextCompact/index.js
@@ -1,7 +1,1 @@
-const optionAlternateTextCompact = {
-    flexShrink: 1,
-    flexGrow: 1,
-    flexBasis: 'auto',
-};
-
-export default optionAlternateTextCompact;
+export {default} from './base';
